requirements.txt
pip~=20.2.2
vine~=1.3.0
amqp~=2.6.1
pytz~=2020.1
PyYAML~=5.3.1
kombu~=4.6.11
billiard~=3.6.3.0
redis~=3.5.3
celery~=4.4.7
certifi~=2020.6.20
chardet~=3.0.4
urllib3~=1.25.10
psycopg2~=2.8.5
idna~=2.10
setuptools~=50.2.0
requests~=2.24.0
sentry-sdk~=0.17.3
kafka-python~=2.0.1